cleaning up nets
