+++ Readme.md
@@ -39,15 +39,15 @@ npm i capi
 Create a `nets.ts` and specify the chains with which you'd like to interact.
 
 ```ts
-import { bins, dev, ws } from "capi"
+import { bins, net } from "capi"
 
 const bin = bins({ polkadot: ["polkadot", "v0.9.38"] })
 
 // The Polkadot relay chain
-export const polkadot = ws({ url: "wss://rpc.polkadot.io/" })
+export const polkadot = net.ws({ url: "wss://rpc.polkadot.io/" })
 
 // A Polkadot development network
-export const polkadotDev = dev({
+export const polkadotDev = net.dev({
   bin: bin.polkadot,
   chain: "polkadot-dev",
 })
