+++ nets.ts
@@ -1,4 +1,4 @@
-import { bins, dev, metadata, ws } from "./mod.ts"
+import { bins, net } from "./mod.ts"
 
 const bin = bins({
   polkadot: ["polkadot", "v0.9.38"],
@@ -9,28 +9,28 @@ const bin = bins({
   trappist: ["trappist-collator", "79bba6e"],
 })
 
-export const polkadotDev = dev({
+export const polkadotDev = net.dev({
   bin: bin.polkadot,
   chain: "polkadot-dev",
 })
-export const polkadot = ws({
+export const polkadot = net.ws({
   url: "wss://rpc.polkadot.io/",
   version: "v0.9.40",
 })
 
-export const westend = ws({ url: "wss://westend-rpc.polkadot.io/" })
-export const statemint = ws({ url: "wss://statemint-rpc.polkadot.io/" })
+export const westend = net.ws({ url: "wss://westend-rpc.polkadot.io/" })
+export const statemint = net.ws({ url: "wss://statemint-rpc.polkadot.io/" })
 
-export const westendDev = dev({
+export const westendDev = net.dev({
   bin: bin.polkadot,
   chain: "westend-dev",
 })
-export const contractsDev = dev({
+export const contractsDev = net.dev({
   bin: bin.substrateContractsNode,
   chain: "dev",
 })
 
-export const rococoDev = dev({
+export const rococoDev = net.dev({
   bin: bin.polkadot,
   chain: "rococo-local",
 })
@@ -45,7 +45,7 @@ export const rococoDevContracts = rococoDev.parachain({
   id: 2000,
 })
 
-export const rococoDevXcm = dev({
+export const rococoDevXcm = net.dev({
   bin: bin.trappistPolkadot,
   chain: "rococo-local",
 })
@@ -60,6 +60,6 @@ export const rococoDevXcmStatemine = rococoDevXcm.parachain({
   id: 1000,
 })
 
-export const polkadotFromMetadata = metadata({
+export const polkadotFromMetadata = net.metadata({
   metadata: await Deno.readFile("examples/raw_rpc/metadata"),
 })
