+++ nets/mod.ts
@@ -1,3 +1,20 @@
+import { DevNetProps } from "./DevNetSpec.ts"
+import { DevRelaySpec } from "./DevRelaySpec.ts"
+import { MetadataNetProps, MetadataNetSpec } from "./MetadataNetSpec.ts"
+import { WsNetProps, WsNetSpec } from "./WsNetSpec.ts"
+
+export namespace net {
+  export function dev(props: DevNetProps) {
+    return new DevRelaySpec(props)
+  }
+  export function ws(props: WsNetProps) {
+    return new WsNetSpec(props)
+  }
+  export function metadata(props: MetadataNetProps) {
+    return new MetadataNetSpec(props)
+  }
+}
+
 // moderate
 
 export * from "./bins.ts"
